+++ src/r0/voip/get_turn_server_info.rs
@@ -1,5 +1,6 @@
 //! [GET /_matrix/client/r0/voip/turnServer](https://matrix.org/docs/spec/client_server/r0.4.0.html#get-matrix-client-r0-voip-turnserver)
 
+use js_int::UInt;
 use ruma_api_macros::ruma_api;
 use serde::{Deserialize, Serialize};
 
@@ -19,7 +20,7 @@ ruma_api! {
         /// The password to use.
         pub password: String,
         /// The time-to-live in seconds.
-        pub ttl: u64,
+        pub ttl: UInt,
         /// A list of TURN URIs.
         pub uris: Vec<String>,
         /// The username to use.
